[Tools][Translator] Add AMD backend support for Triton-to-Gluon translator (#9717)
Extends the existing NVIDIA Triton-to-Gluon translator to support AMD
GPU architectures, with gfx1250 as the primary target and CDNA3/CDNA4
support for non-descriptor ops.

### Changes

- **Added AMD backend to translator helpers**: Translates Triton ops to
Gluon equivalents using AMD-specific hardware features -- WMMA for
`tl.dot`, TDM for tensor descriptors (`tl.make_tensor_descriptor`,
`desc.load`, `desc.store`), and `PaddedSharedLayout` for shared memory.
- **TDM gather/scatter** (`translator_helpers.py`): Implements
`tl_obj_gather_amd` / `tl_obj_scatter_amd` as `@builtin` helpers that
recreate the TDM descriptor with the correct block shape for multi-row
gather/scatter operations.
- **CDNA3/CDNA4 support** (`translator_helpers.py`): Adds MFMA-based
`tl.dot` translation, 64-thread-per-warp layouts, and target-aware
dispatch for gfx942/gfx950.
- **Target propagation** (`translator.py`, `translator_helpers.py`):
Injects `helpers._current_target = helpers._make_target("{target}")`
into generated code for non-NVIDIA targets. Architecture detection
helpers (`get_num_threads_per_warp`, `_is_gfx1250`, `_is_cdna`, etc.)
accept an optional `target` parameter with fallback to
`current_target()`.
- **Host descriptor support** (`inline_helpers.py`): Adds
`convert_host_descriptor_amd` using `gluon.amd.gfx1250.TensorDescriptor`
with `PaddedSharedLayout`. Deduplicates shared `torch_dtype_to_triton`
helper.
- **Segfault fix** (`ir.cc`): Uses `isa_and_nonnull` /
`dyn_cast_or_null` in `getTensorDescMetadata` to handle unencoded tensor
descriptor block types from the `make_ttgir` path. This was needed in
order to make the reference triton tensor descriptor tests pass, where
the descriptors were created on the host side.
- **Tests** (`test_triton_to_gluon.py`): Parametrizes existing tests
across `nvidia`/`gfx1250`/`gfx942`/`gfx950` targets. Adds roundtrip
tests for `tl.cat`, `tl.make_tensor_descriptor`, and gather/scatter.

### Known limitation

`tl.make_tensor_descriptor` / `desc.load` / `desc.store` translation is
not yet supported for CDNA3/CDNA4 targets. The Gluon builder lacks
generic `DescriptorLoad`/`DescriptorStore` ops, so descriptor-based
kernels currently only translate for gfx1250 (TDM) and NVIDIA (TMA).
